Rename errors.py to exceptions.py to be pedantic
