+++ smolmodels/exceptions.py
